Fix: New Post Modal lazy loading
performance fixes
