+++ src/AppRouter.tsx
@@ -1,9 +1,19 @@
+import { lazy, Suspense } from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
 
 import { PostDetail } from './components/molecules/PostDetail/PostDetail';
-import { Albums } from './components/organisms/Albums/Albums';
-import { Posts } from './components/organisms/Posts/Posts';
-import { Todos } from './components/organisms/Todos/Todos';
+import Posts from './components/organisms/Posts/Posts';
+
+const Albums = lazy(() =>
+  import('./components/organisms/Albums/Albums').then((module) => ({
+    default: module.default
+  }))
+);
+const Todos = lazy(() =>
+  import('./components/organisms/Todos/Todos').then((module) => ({
+    default: module.default
+  }))
+);
 
 export const AppRouter = () => {
   return (
@@ -14,8 +24,22 @@ export const AppRouter = () => {
         <Route path=":postId" element={<PostDetail />} />
       </Route>
 
-      <Route path="albums" element={<Albums />} />
-      <Route path="todos" element={<Todos />} />
+      <Route
+        path="albums"
+        element={
+          <Suspense>
+            <Albums />
+          </Suspense>
+        }
+      />
+      <Route
+        path="todos"
+        element={
+          <Suspense>
+            <Todos />
+          </Suspense>
+        }
+      />
     </Routes>
   );
 };
